requirements.txt
# Local wheels for packages not on public PyPI
--find-links wheels
--pre

# Core — Microsoft Agent Framework (pinned to bundled wheels)
agent-framework-core==1.0.0rc6
agent-framework-azure-ai==1.0.0rc6
botbuilder-core>=4.16.0
aiohttp>=3.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
tabulate>=0.9.0
pyodbc>=5.0.0  # Required when DATASOURCE=sql; needs ODBC Driver 17 or 18 installed on the host
